Separated out concerns for main menu
