--- src/components/Menu/index.js
@@ -1,3 +1,0 @@
-export { default as Timeline } from './Timeline';
-export { default as TimelineStage } from './TimelineStage';
-export { default as StatsPanel } from './StatsPanel';
