+++ src/containers/MainMenu/StagesMenu.js
@@ -1,0 +1,14 @@
+import { connect } from 'react-redux';
+import StagesMenu from '../../components/MainMenu/StagesMenu';
+import { stages } from '../../selectors/session';
+
+function mapStateToProps(state) {
+  const currentStages = stages(state);
+
+  return {
+    currentStages,
+  };
+}
+
+export default connect(mapStateToProps)(StagesMenu);
+
