--- src/containers/Menu/index.js
@@ -1,3 +1,0 @@
-export { default as MenuContainer } from './MenuContainer';
-export { default as SettingsMenu } from './SettingsMenu';
-export { default as StagesMenu } from './StagesMenu';
